+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/impl/SmkUseImpl.kt
@@ -9,8 +9,8 @@ import com.jetbrains.python.psi.PyElementVisitor
 import com.jetbrains.python.psi.types.TypeEvalContext
 import com.jetbrains.snakecharm.lang.parser.SmkTokenTypes
 import com.jetbrains.snakecharm.lang.psi.*
-import com.jetbrains.snakecharm.lang.psi.elementTypes.SmkElementTypes
 import com.jetbrains.snakecharm.lang.psi.elementTypes.SmkElementTypes.USE_IMPORTED_RULES_NAMES
+import com.jetbrains.snakecharm.lang.psi.elementTypes.SmkElementTypes.USE_NEW_NAME_PATTERN
 import com.jetbrains.snakecharm.lang.psi.elementTypes.SmkStubElementTypes
 import com.jetbrains.snakecharm.lang.psi.stubs.SmkUseStub
 import com.jetbrains.snakecharm.lang.psi.types.SmkRuleLikeSectionType
@@ -34,14 +34,9 @@ class SmkUseImpl : SmkRuleLikeImpl<SmkUseStub, SmkUse, SmkRuleOrCheckpointArgsSe
     }
 
     override fun getNameNode(): ASTNode? {
-        val identifier = super.getNameNode()
-        if (identifier != null) { // Returns new name if we know it
+        val namePattern = getNewNamePattern()
+        if (namePattern != null) { // Returns name identifier if it exits
             // Example: use rule A as new_A
-            // Here we can detect name node by default
-            return identifier
-        }
-        val namePattern = getNameIdentifierPattern()
-        if (namePattern != null) { // Returns name patter if it exits
             // Example: use rule A, B from M as new_*
             // There are pattern instead of single node
             return namePattern.node
@@ -60,18 +55,18 @@ class SmkUseImpl : SmkRuleLikeImpl<SmkUseStub, SmkUse, SmkRuleOrCheckpointArgsSe
     }
 
     override fun getProducedRulesNames(visitedFiles: MutableSet<PsiFile>): List<Pair<String, PsiElement>> {
-        val identifier = super.getNameNode()
-        if (identifier != null) {
-            return listOf(identifier.text to identifier.psi)
+        val newName = getNewNamePattern()
+
+        if (newName != null && !newName.isWildcard()) {
+            return listOf(newName.text to newName.originalElement)
         }
-        val newName = getNameIdentifierPattern()
         val originalNames = mutableListOf<Pair<String, PsiElement>>()
         getDefinedReferencesOfImportedRuleNames()?.forEach { reference ->
             originalNames.add(reference.text to reference)
         }
         if (originalNames.isEmpty()) {
             val pairs = getPairsOfImportedRulesAndNames(visitedFiles) ?: return emptyList()
-            originalNames.addAll(pairs.map { it.first to it.second })
+            originalNames.addAll(pairs)
         }
         return if (newName != null) {
             originalNames.map { newName.text.replace("*", it.first) to it.second }
@@ -80,36 +75,31 @@ class SmkUseImpl : SmkRuleLikeImpl<SmkUseStub, SmkUse, SmkRuleOrCheckpointArgsSe
         }
     }
 
-    /**
-     * Returns a [PsiElement] which sets pattern of produced rule names
-     */
-    private fun getNameIdentifierPattern() = findChildByType(SmkElementTypes.USE_NAME_IDENTIFIER) as? PsiElement
+    override fun getNewNamePattern() = findChildByType(USE_NEW_NAME_PATTERN) as? SmkUseNewNamePattern
+    override fun getImportedNamesList() = findChildByType(USE_IMPORTED_RULES_NAMES) as? SmkImportedRulesNamesList
 
     /**
      * Collects all imported rules and their names
      */
     private fun getPairsOfImportedRulesAndNames(visitedFiles: MutableSet<PsiFile>): List<Pair<String, SmkRuleOrCheckpoint>>? {
         val module = ((getModuleName() as? SmkReferenceExpression)?.reference?.resolve() as? SmkModule) ?: return null
-        val file = module.getPsiFile()
-        return if (file != null && file is SmkFile) {
-            file.advancedCollectRules(visitedFiles).map { it.first to it.second }
-        } else {
-            null
+
+        return when (val psiFile = module.getPsiFile()) {
+            is SmkFile -> psiFile.advancedCollectRules(visitedFiles).map { it.first to it.second }
+            else -> null
         }
     }
 
     override fun getModuleName() =
         (findChildByType(SmkTokenTypes.SMK_FROM_KEYWORD) as? PsiElement)?.nextSibling?.nextSibling
 
-    override fun getDefinedReferencesOfImportedRuleNames(): Array<SmkReferenceExpression>? = PsiTreeUtil.getChildrenOfType(
-        findChildByType(USE_IMPORTED_RULES_NAMES),
-        SmkReferenceExpression::class.java
-    )
+    override fun getDefinedReferencesOfImportedRuleNames(): Array<SmkReferenceExpression>? =
+        PsiTreeUtil.getChildrenOfType(
+            findChildByType(USE_IMPORTED_RULES_NAMES),
+            SmkReferenceExpression::class.java
+        )
 
     override fun getImportedRules(): List<SmkRuleOrCheckpoint>? =
-        getPairsOfImportedRulesAndNames(mutableSetOf())?.map { it.second }
-
-    override fun nameIdentifierIsWildcard() = nameIdentifier?.let {
-        it is SmkUseNameIdentifier && it.textContains('*')
-    } ?: false
+        getDefinedReferencesOfImportedRuleNames()?.mapNotNull { it.reference.resolve() as? SmkRuleOrCheckpoint }
+            ?: getPairsOfImportedRulesAndNames(mutableSetOf())?.map { it.second }
 }