+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/impl/refs/SmkRuleOrCheckpointNameReference.kt
@@ -16,7 +16,7 @@ import com.jetbrains.snakecharm.lang.psi.SmkFile
 import com.jetbrains.snakecharm.lang.psi.SmkModule
 import com.jetbrains.snakecharm.lang.psi.SmkReferenceExpression
 import com.jetbrains.snakecharm.lang.psi.SmkUse
-import com.jetbrains.snakecharm.lang.psi.SmkImportedRulesNames
+import com.jetbrains.snakecharm.lang.psi.SmkImportedRulesNamesList
 import com.jetbrains.snakecharm.lang.psi.stubs.SmkModuleNameIndex
 import com.jetbrains.snakecharm.lang.psi.types.SmkCheckpointType
 import com.jetbrains.snakecharm.lang.psi.types.SmkRulesType
@@ -56,7 +56,7 @@ class SmkRuleOrCheckpointNameReference(
 
         val moduleRef = element.parentOfType<SmkUse>()?.getModuleName() as? SmkReferenceExpression
         val itIsModuleMameReference = element.parent is SmkUse
-        val parentIsImportedRuleNames = element.parent is SmkImportedRulesNames
+        val parentIsImportedRuleNames = element.parent is SmkImportedRulesNamesList
         if (!parentIsImportedRuleNames && !itIsModuleMameReference) {
             return results
         }